Loading TestPointCloud from File and displaying it * Starting to handle Camera

+++ FaceScanKinect/src/KinectGrabber.cpp
@@ -248,7 +248,7 @@ bool KinectGrabber::CreatePointCloud() {
                             pointCloudColors.push_back(RGB3f(&rgbx.rgbBlue));
 
                             Q_ASSERT(pointCloudColors.size() == pointCloudPoints.size());
-                            emit PointCloudDataAvailable(&pointCloudPoints[0], &pointCloudColors[0], pointCloudPoints.size());
+                            emit PointCloudDataAvailable(&pointCloudPoints[0], &pointCloudColors[0], (int)pointCloudPoints.size());
                         }
                     }
                 }
